backend_django/requirements.txt
Django>=4.2
djangorestframework
gunicorn
psycopg2-binary
scikit-learn
joblib
numpy
whitenoise
